+++ src/clua-machine.cpp
@@ -71,7 +71,7 @@ static int machine_class_index_verify_reset_uarch(lua_State *L) {
 static int machine_class_index_verify_send_cmio_response(lua_State *L) {
     lua_settop(L, 6);
     const uint16_t reason = static_cast<uint16_t>(luaL_checkinteger(L, 1));
-    uint64_t length{0};
+    size_t length{0};
     // NOLINTNEXTLINE(cppcoreguidelines-pro-type-reinterpret-cast)
     const auto *data = reinterpret_cast<const unsigned char *>(luaL_checklstring(L, 2, &length));
     const char *log = clua_check_schemed_json_string(L, 4, "AccessLog");
